confirmation: Move check_prereg_key_and_redirect to registration.py.
